Charly/requirements.txt
fastapi
uvicorn[standard]
python-multipart
mistralai
google-genai
chromadb
pdfplumber
pdf2image
Pillow
python-dotenv
